% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{annotate_gct}
\alias{annotate_gct}
\title{Add annotations to a GCT_object}
\usage{
annotate_gct(ds, annot, dimension = "column", keyfield = "id")
}
\arguments{
\item{ds}{a GCT_object}

\item{annot}{a \code{\link{data.frame}} of annotations}

\item{dimension}{either 'row' or 'column' indicating which dimension
of \code{g} to annotate}

\item{keyfield}{the character name of the column in \code{annot} that 
matches the row or column identifiers in \code{g}}
}
\value{
a GCT_object with annotations applied to the specified
  dimension
}
\description{
Given a GCT_object and either a \code{\link{data.frame}} or
a path to an annotation table, apply the annotations to the
gct using the given \code{keyfield}.
}
\examples{
#Add column annotations
gct_file <- system.file("extdata", "example_n50x100.gct", package="mapGCT")
(ds <- parse_gct(gct_file))
ds <- annotate_gct(ds, col_annotate)

}
\concept{GCT utilities}
