% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_annotate}
\alias{col_annotate}
\title{An example table of column annotations, which is to be 
annotated to existing gct_object.}
\format{An object of class \code{data.frame} with 50 rows and 3 columns.}
\usage{
col_annotate
}
\description{
An example table of column annotations, which is to be 
annotated to existing gct_object.
}
\keyword{datasets}
