% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_desc}
\alias{col_desc}
\title{An example table of column descriptions, as would be
parsed from or parse_gct. Initially all the
columns are of type character.}
\format{An object of class \code{data.frame} with 50 rows and 4 columns.}
\usage{
col_desc
}
\description{
An example table of column descriptions, as would be
parsed from or parse_gct. Initially all the
columns are of type character.
}
\keyword{datasets}
