% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gct_mat}
\alias{gct_mat}
\title{An example of numeric matrix, as would be
parsed from parse_gct or used to create gct_object
by to_GCT}
\format{An object of class \code{matrix} with 100 rows and 50 columns.}
\usage{
gct_mat
}
\description{
An example of numeric matrix, as would be
parsed from parse_gct or used to create gct_object
by to_GCT
}
\keyword{datasets}
