% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{GCT_object-class}
\alias{GCT_object-class}
\title{An S4 class to represent a GCT object}
\description{
The GCT_object class serves to represent annotated matrices.
The \code{mat} slot contains numeric matrics and the \code{rdesc} and \code{cdesc}
slots contain data frames with annotations about the rows and columns, respectively
}
\section{Slots}{

\describe{
\item{\code{mat}}{a numeric matrix}

\item{\code{rid}}{a character vector of row ids}

\item{\code{cid}}{a character vector of column ids}

\item{\code{rdesc}}{a \code{data.frame} of row descriptors}

\item{\code{cdesc}}{a \code{data.frame} of column descriptors}

\item{\code{version}}{version of the gct}
}}

