% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{row_desc}
\alias{row_desc}
\title{An example table of row descriptons, as would be
parsed from parse_gct or used to create gct_object
by to_GCT}
\format{An object of class \code{data.frame} with 100 rows and 3 columns.}
\usage{
row_desc
}
\description{
An example table of row descriptons, as would be
parsed from parse_gct or used to create gct_object
by to_GCT
}
\keyword{datasets}
