% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_gct}
\alias{write_gct}
\title{Write a GCT_object to disk in GCT format}
\usage{
write_gct(ds, ofile, precision = 4, ver = 3)
}
\arguments{
\item{ds}{the GCT_object to be written}

\item{ofile}{the desired output filename}

\item{precision}{the numeric precision at which to
save the matrix. See \code{details}.}

\item{ver}{the GCT version to write. See \code{details}.}

\item{appenddim}{boolean indicating whether to append
matrix dimensions to filename}
}
\description{
Write a GCT_object to disk in GCT format
}
\details{
Since GCT is text format, the higher \code{precision}
  you choose, the larger the file size.
  \code{ver} is assumed to be 3, aka GCT version 1.3, which supports
  embedded row and column metadata in the GCT file. Any other value
  passed to \code{ver} will result in a GCT version 1.2 file which
  contains only the matrix data and no annotations.
}
\examples{
ds <- to_GCT(mat = gct_mat, cdesc = col_desc, rdesc = row_desc)
write_gct(ds, "dataset.gct", precision=2)

}
\seealso{
Other GCT parsing functions: \code{\link{parse_gct}}
}
