% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parse_gct}
\alias{parse_gct}
\title{Parse a GCT file into the workspace as a GCT_object}
\usage{
parse_gct(fname, matrix_only = F)
}
\arguments{
\item{fname}{path to the GCT file on disk}

\item{matrix_only}{boolean indicating whether to parse only
the matrix (ignoring row and column annotations)}
}
\description{
Parse a GCT file into the workspace as a GCT_object
}
\details{
\code{parse_gct} also supports parsing of plain text
  GCT files, so this function can be used as a general GCT parser.
}
\examples{
gct_file <- system.file("extdata", "example_n50x100.gct", package="mapGCT")
(ds <- parse_gct(gct_file))

# matrix only
ds <- parse_gct(gct_file, matrix_only=TRUE)

}
\seealso{
Other GCT parsing functions: 
\code{\link{write_gct}()}
}
\concept{GCT parsing functions}
