% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{to_GCT}
\alias{to_GCT}
\title{Create a GCT_object for a given numeric matrix}
\usage{
to_GCT(mat, cdesc = NULL, rdesc = NULL, version = NULL)
}
\arguments{
\item{mat}{a numeric matrix}

\item{cdesc}{a \code{data.frame} of column descriptors}

\item{rdesc}{a \code{data.frame} of row descriptors}

\item{version}{version of the gct}
}
\description{
create a GCT_object for a given matrix and cloumn description
}
